+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/mods/EclipticSeasonsHelper.java
@@ -90,4 +90,24 @@ public class EclipticSeasonsHelper implements ISeasonModHelper {
 
     return duration;
   }
+
+  @Override
+  public boolean infertileBiome(Player player) {
+    return false;
+  }
+
+  @Override
+  public boolean alwaysWinterBiome(Player player) {
+    return false;
+  }
+
+  @Override
+  public boolean undergroundFertility(Player player) {
+    return true;
+  }
+
+  @Override
+  public boolean fertileBiome(Player player) {
+    return true;
+  }
 }