+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/mods/TerrafirmaCraftHelper.java
@@ -149,6 +149,26 @@ public class TerrafirmaCraftHelper implements ISeasonModHelper {
     }
   }
 
+  @Override
+  public boolean infertileBiome(Player player) {
+    return false;
+  }
+
+  @Override
+  public boolean alwaysWinterBiome(Player player) {
+    return false;
+  }
+
+  @Override
+  public boolean undergroundFertility(Player player) {
+    return true;
+  }
+
+  @Override
+  public boolean fertileBiome(Player player) {
+    return true;
+  }
+
   /**
    * <h1>TerrafirmaCraft Seasons</h1>
    * <pre>